backend/pyproject.toml
[project]
name = "pokeparse-backend"
version = "0.1.0"
description = "Localhost backend for PokeParse: AI card analysis + Telegram alerts"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0,<4",
    "openai>=1.50,<2",
    "python-dotenv>=1.0,<2",
    "requests>=2.31,<3",
]

[tool.uv]
exclude-newer = "30 days"
